Python API/JWT Studies/requirements.txt
fastapi
uvicorn
pyjwt
PyJWT